requirements.txt
ultralytics
opencv-python
datasets
tqdm
pyyaml
